LayoutTests/http/tests/history/cross-origin-replace-history-object-child-expected.txt

CONSOLE MESSAGE: line 30: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
ALERT: Child window's history object before attempt to clear: undefined
ALERT: About to shadow child window's history object: [object History]
ALERT: PASS: Could not shadow child window's history object: [object History]
CONSOLE MESSAGE: line 18: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
ALERT: Child window's history object after attempt to clear: undefined
